Started Work On First Game Level
Started working on the first game level. I've also made a few improvements here and there. Currently all the code for the first level is in levelOne.js. I'm positioning level parts based on percents of the the canvas width and height. The idea is that the game size will scale if the canvas size changes.