I Have The Player Drawing To Screen
It took a lot of tweaking percents and offsets but I finally have the player and all it's parts drawing to the screen. The sizes of all the player parts scale according to the height and width allocated for the player. The best looking results are when there is a 2:1 ratio bewteen height and width of the player.