Big Success! Player can now climb ladders
After a lot of work, the player can now climb ladders. The mechanics allow for the player to move up and down on the Y axis once on a ladder but cannot jump until the players head is at the top of the ladder. all jumps are ended once on a ladder. This took so much work the feels great to play with!